refactor(fileUtils): move fileUtils to config folder
